Revert "Feature/support_for_saving_and_loading_from_file"
