+++ src/main/java/module-info.java
@@ -3,7 +3,9 @@ module org.group2.textprocessingtool {
     requires javafx.fxml;
 
 
-    opens org.group2.textprocessingtool.controllers to javafx.fxml;
-    exports org.group2.textprocessingtool;
+    opens org.group2.textprocessingtool to javafx.fxml;
+    exports org.group2.textprocessingtool.controllers;
 
+    // Add the following line to export the package to javafx.graphics
+    exports org.group2.textprocessingtool to javafx.graphics;
 }
