+++ src/main/java/org/group2/textprocessingtool/controllers/EditMenuController.java
@@ -1,0 +1,27 @@
+package org.group2.textprocessingtool.controllers;
+
+import javafx.event.ActionEvent;
+
+public class EditMenuController {
+    public void handleUndo(ActionEvent actionEvent) {
+    }
+
+    public void handleRedo(ActionEvent actionEvent) {
+    }
+
+    public void handleCut(ActionEvent actionEvent) {
+    }
+
+    public void handleCopy(ActionEvent actionEvent) {
+    }
+
+    public void handlePaste(ActionEvent actionEvent) {
+    }
+
+    public void handleFind(ActionEvent actionEvent) {
+    }
+
+    public void handleReplace(ActionEvent actionEvent) {
+
+    }
+}
