+++ src/main/java/org/group2/textprocessingtool/controllers/FileMenuController.java
@@ -1,0 +1,18 @@
+package org.group2.textprocessingtool.controllers;
+
+import javafx.event.ActionEvent;
+
+public class FileMenuController {
+    public void handleOpen(ActionEvent actionEvent) {
+    }
+
+    public void handleSave(ActionEvent actionEvent) {
+    }
+
+    public void handleSaveAs(ActionEvent actionEvent) {
+    }
+
+    public void handleExit(ActionEvent actionEvent) {
+
+    }
+}
