+++ src/main/java/org/group2/textprocessingtool/controllers/HelpMenuController.java
@@ -1,0 +1,9 @@
+package org.group2.textprocessingtool.controllers;
+
+import javafx.event.ActionEvent;
+
+public class HelpMenuController {
+    public void handleAbout(ActionEvent actionEvent) {
+
+    }
+}
