+++ src/main/java/org/group2/textprocessingtool/controllers/RegexMenuController.java
@@ -1,0 +1,21 @@
+package org.group2.textprocessingtool.controllers;
+
+import javafx.event.ActionEvent;
+
+public class RegexMenuController {
+    public void handleCustomRegex(ActionEvent actionEvent) {
+    }
+
+    public void handleReplaceWithRegex(ActionEvent actionEvent) {
+    }
+
+    public void handleMatchEmail(ActionEvent actionEvent) {
+    }
+
+    public void handleMatchPhone(ActionEvent actionEvent) {
+    }
+
+    public void handleMatchDate(ActionEvent actionEvent) {
+
+    }
+}
