+++ src/main/resources/org/group2/textprocessingtool/EditMenu.fxml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Menu?>
+<?import javafx.scene.control.MenuItem?>
+
+<Menu xmlns:fx="http://javafx.com/fxml" text="Edit" fx:controller="org.group2.textprocessingtool.controllers.EditMenuController">
+    <MenuItem text="Undo" onAction="#handleUndo"/>
+    <MenuItem text="Redo" onAction="#handleRedo"/>
+    <MenuItem text="Cut" onAction="#handleCut"/>
+    <MenuItem text="Copy" onAction="#handleCopy"/>
+    <MenuItem text="Paste" onAction="#handlePaste"/>
+    <MenuItem text="Find" onAction="#handleFind"/>
+    <MenuItem text="Replace" onAction="#handleReplace"/>
+</Menu>
