+++ src/main/resources/org/group2/textprocessingtool/FileMenu.fxml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Menu?>
+<?import javafx.scene.control.MenuItem?>
+
+<Menu xmlns:fx="http://javafx.com/fxml" text="File" fx:controller="org.group2.textprocessingtool.controllers.FileMenuController">
+    <MenuItem text="Open" onAction="#handleOpen"/>
+    <MenuItem text="Save" onAction="#handleSave"/>
+    <MenuItem text="Save As" onAction="#handleSaveAs"/>
+    <MenuItem text="Exit" onAction="#handleExit"/>
+</Menu>
