+++ src/main/resources/org/group2/textprocessingtool/HelpMenu.fxml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Menu?>
+<?import javafx.scene.control.MenuItem?>
+
+<Menu xmlns:fx="http://javafx.com/fxml" text="Help" fx:controller="org.group2.textprocessingtool.controllers.HelpMenuController">
+    <MenuItem text="About" onAction="#handleAbout"/>
+</Menu>
