+++ src/main/resources/org/group2/textprocessingtool/RegexMenu.fxml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Menu?>
+<?import javafx.scene.control.MenuItem?>
+
+<Menu xmlns:fx="http://javafx.com/fxml" text="Regex" fx:controller="org.group2.textprocessingtool.controllers.RegexMenuController">
+    <MenuItem text="Custom regex" onAction="#handleCustomRegex"/>
+    <MenuItem text="Replace with regex" onAction="#handleReplaceWithRegex"/>
+    <MenuItem text="Match e-mail" onAction="#handleMatchEmail"/>
+    <MenuItem text="Match phone" onAction="#handleMatchPhone"/>
+    <MenuItem text="Match Date" onAction="#handleMatchDate"/>
+</Menu>
